Add nats.aio.msg to avoid circular imports
Make Srv a dataclass
