+++ nats/js/client.py
@@ -17,10 +17,11 @@ import json
 import time
 import nats.errors
 import nats.js.errors
+from nats.aio.msg import Msg
 from nats.aio.subscription import Subscription
 from nats.js.manager import JetStreamManager
 from nats.js import api
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional, Callable
 from dataclasses import asdict
 
 LAST_CONSUMER_SEQ_HDR = "Nats-Last-Consumer"
@@ -107,7 +108,7 @@ class JetStream:
         self,
         subject: str,
         queue: Optional[str] = None,
-        cb=None,
+        cb: Optional[Callable[['Msg'], None]] = None,
         durable: Optional[str] = None,
         stream: Optional[str] = None,
         config: Optional[api.ConsumerConfig] = None,
